Test 브랜치 작업 내용을 Main 브랜치로 병합합니다.

+++ src/main/java/com/challenger/challengerbe/modules/challenge/controller/ChallengeUserController.java
@@ -2,7 +2,10 @@ package com.challenger.challengerbe.modules.challenge.controller;
 
 import com.challenger.challengerbe.auth.login.AuthInfo;
 import com.challenger.challengerbe.common.CommonResponse;
+import com.challenger.challengerbe.common.exception.AlreadyExistException;
+import com.challenger.challengerbe.common.exception.AlreadyUseException;
 import com.challenger.challengerbe.modules.challenge.dto.*;
+import com.challenger.challengerbe.modules.challenge.service.ChallengeApplyService;
 import com.challenger.challengerbe.modules.challenge.service.ChallengeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -39,12 +42,15 @@ public class ChallengeUserController {
 
     private final ChallengeService challengeService;
 
+    private final ChallengeApplyService challengeApplyService;
+
     @Operation(summary = "챌린지 목록 조회(페이징 포함)")
     @Parameters({
             @Parameter(name = "code", description = "키워드 (만약 전체일 경우 빈값으로 전달 주시면 되겠습니다.",required = false),
             @Parameter(name = "startDate", description = "시작일자 (string 타입으로 전달 주시면 됩니다.)", example = "0000-00-00",required = false),
             @Parameter(name = "endDate", description = "마지막일자 (string 타입으로 전달 주시면 됩니다.)", example = "0000-00-00",required = false),
-            @Parameter(name = "activeStatus", description = "진행상태 ", example = "ex) JOIN(모집중), KEEP(진행중), END(종료) ",required = false)
+            @Parameter(name = "activeStatus", description = "진행상태 ", example = "ex) JOIN(모집중), KEEP(진행중), END(종료) ",required = false),
+            @Parameter(name = "page", description = "페이지 번호 ", example = "1",required = false)
     })
     @GetMapping("/challenge/list")
     public Page<ChallengeSummaryResponse> selectChallengeList(@Parameter(hidden = true) ChallengeDefaultDto searchDto,
@@ -83,8 +89,8 @@ public class ChallengeUserController {
     @Parameters({
             @Parameter(name = "_file", description = "formData 에서 파일 형식 ,파일 이름에 해당 , 파일명은 아무거나 선언가능 _file은 예시"),
             @Parameter(name = "_alt_file", description = "formData 에서 string, 파일의 비고명 아무거나 선언가능 대신 파일명의 name을 뒤에 그대로 붙여줘야함 _alt{파일name}"),
-            @Parameter(name = "_modify_file", description = "수정하는 파일에 대한 parent_idx 값 (필수로 들어가야함), 파일의 비고명 아무거나 선언가능 대신 파일명의 name을 뒤에 그대로 붙여줘야함 _modify{파일name}",required = true),
-            @Parameter(name = "_delete_file", description = "삭제하는 파일에 대한 parent_idx 값, 파일의 비고명 아무거나 선언가능 대신 파일명의 name을 뒤에 그대로 붙여줘야함 _delete{파일name}")
+            @Parameter(name = "_modify_file", description = "file idx 값 (필수로 들어가야함), 파일의 비고명 아무거나 선언가능 대신 파일명의 name을 뒤에 그대로 붙여줘야함 _modify{파일name}",required = true),
+            @Parameter(name = "_delete_file", description = "file idx 값 (이미지와 상관없이 기존에 존재하는 이미지를 지우고자 할 경우 선택), 파일의 비고명 아무거나 선언가능 대신 파일명의 name을 뒤에 그대로 붙여줘야함 _delete{파일name}")
     })
     @ApiResponses(value={
             @ApiResponse(responseCode = "200", description = "수정 완료"),
@@ -94,14 +100,9 @@ public class ChallengeUserController {
     public ResponseEntity<?> updateChallenge(final @Valid @RequestPart ChallengeUpdateRequest challengeUpdateRequest,
                                              @Parameter(hidden = true) @AuthInfo String token) throws Exception {
 
-        try{
-            ChallengeDto challengeDto = ChallengeDto.updateOf(challengeUpdateRequest);
-            challengeDto.setIdk(token);
-            challengeService.updateChallenge(challengeDto);
-        }catch (Exception e) {
-            logger.error("update challenge info error : {}",e.getMessage());
-            return new ResponseEntity<>(CommonResponse.resOnlyMessageOf("수정시 오류가 발생했습니다."), HttpStatus.BAD_REQUEST);
-        }
+        ChallengeDto challengeDto = ChallengeDto.updateOf(challengeUpdateRequest);
+        challengeDto.setIdk(token);
+        challengeService.updateChallenge(challengeDto);
 
         return new ResponseEntity<>(CommonResponse.resOnlyMessageOf("수정 되었습니다."),HttpStatus.OK);
     }
@@ -114,15 +115,15 @@ public class ChallengeUserController {
     })
     @DeleteMapping("/challenge/del")
     public ResponseEntity<?> deleteChallenge(@RequestParam("idx") Long idx) throws Exception {
-        
-        try{
-            ChallengeDto challengeDto = new ChallengeDto();
-            challengeDto.setIdx(idx);
-            challengeService.deleteChallenge(challengeDto);
-        }catch (Exception e) {
-            logger.error("delete challenge info error : {}",e.getMessage());
-            return new ResponseEntity<>(CommonResponse.resOnlyMessageOf("삭제시 오류가 발생했습니다."), HttpStatus.BAD_REQUEST);
+
+        int totCnt = challengeApplyService.selectChallengeUserTotalCountByChallengeIdx(idx);
+        if(totCnt > 0) {
+            throw new AlreadyUseException("이미 참여중인 참여자들이 있어 삭제가 불가능합니다.");
         }
+
+        ChallengeDto challengeDto = new ChallengeDto();
+        challengeDto.setIdx(idx);
+        challengeService.deleteChallenge(challengeDto);
 
         return new ResponseEntity<>(CommonResponse.resOnlyMessageOf("삭제 되었습니다."),HttpStatus.OK);
     }
